+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # GCP Tamr Wrapper
 
+## v3.1.0 - April 20th 2023
+* Following optional input variables now available
+  * tamr_instance_tags
+  * dataproc_network_tags
+* Allows flexibility for the user of this module to supply network tags to tamr_vm module and dataproc node
+
 ## v3.0.1 - April 7th 2023
 * Use patched iam module